+++ frontend/app/src/components/asset-manager/managed/ManagedAssetIgnoreSwitch.vue
@@ -1,40 +1,75 @@
 <script setup lang="ts">
+import { useIgnoredAssetsStore } from '@/store/assets/ignored';
 import { useWhitelistedAssetsStore } from '@/store/assets/whitelisted';
 
 const props = withDefaults(
   defineProps<{
     identifier: string;
     isSpam?: boolean;
+    showMoreOptions?: boolean;
+    loadingIgnore?: boolean;
+    loadingWhitelist?: boolean;
+    loadingSpam?: boolean;
   }>(),
   {
     isSpam: false,
+    loadingIgnore: false,
+    loadingSpam: false,
+    loadingWhitelist: false,
+    showMoreOptions: false,
   },
 );
 
 const emit = defineEmits<{
-  (e: 'toggle-whitelist'): void;
-  (e: 'toggle-spam'): void;
+  'toggle-ignore': [];
+  'toggle-whitelist': [];
+  'toggle-spam': [];
 }>();
 
 const { identifier } = toRefs(props);
 
 const { t } = useI18n({ useScope: 'global' });
 
+const { useIsAssetIgnored } = useIgnoredAssetsStore();
 const { isAssetWhitelisted } = useWhitelistedAssetsStore();
+
+const isIgnored = useIsAssetIgnored(identifier);
 const isWhitelisted = isAssetWhitelisted(identifier);
 
-function toggleWhitelist() {
-  emit('toggle-whitelist');
-}
+const isLoading = computed<boolean>(() => props.loadingIgnore || props.loadingWhitelist || props.loadingSpam);
 
-function toggleSpam() {
-  emit('toggle-spam');
-}
+const isMenuLoading = computed<boolean>(() => props.loadingWhitelist || props.loadingSpam);
+
+const isIgnoringDisabled = computed<boolean>(() => props.isSpam || get(isWhitelisted) || get(isLoading));
+
+const tooltipMessage = computed<string>(() =>
+  props.isSpam ? t('ignore.spam.hint') : t('ignore.whitelist.hint'),
+);
 </script>
 
 <template>
-  <div class="flex items-center">
+  <div class="flex justify-start items-center gap-2">
+    <RuiTooltip
+      :popper="{ placement: 'top' }"
+      :open-delay="400"
+      tooltip-class="max-w-[10rem]"
+      :disabled="!isIgnoringDisabled || isLoading"
+    >
+      <template #activator>
+        <RuiSwitch
+          color="primary"
+          hide-details
+          :loading="loadingIgnore"
+          :disabled="isIgnoringDisabled"
+          :model-value="isIgnored"
+          @update:model-value="emit('toggle-ignore')"
+        />
+      </template>
+      {{ tooltipMessage }}
+    </RuiTooltip>
+
     <RuiMenu
+      v-if="showMoreOptions"
       menu-class="w-[15rem]"
       close-on-content-click
     >
@@ -53,6 +88,7 @@ function toggleSpam() {
             icon
             v-bind="attrs"
             size="sm"
+            :loading="isMenuLoading"
             class="dark:!bg-rui-grey-800 dark:!text-white"
           >
             <RuiIcon
@@ -66,7 +102,8 @@ function toggleSpam() {
         <RuiButton
           variant="list"
           size="sm"
-          @click="toggleWhitelist()"
+          :disabled="isLoading"
+          @click="emit('toggle-whitelist')"
         >
           <template #prepend>
             <RuiCheckbox
@@ -81,7 +118,8 @@ function toggleSpam() {
         <RuiButton
           variant="list"
           size="sm"
-          @click="toggleSpam()"
+          :disabled="isLoading"
+          @click="emit('toggle-spam')"
         >
           <template #prepend>
             <RuiCheckbox