+++ frontend/app/src/components/asset-manager/managed/ManagedAssetTable.vue
@@ -5,7 +5,7 @@ import type { Filters, Matcher } from '@/composables/filters/assets';
 import type { Collection } from '@/types/collection';
 import AssetUnderlyingTokens from '@/components/asset-manager/AssetUnderlyingTokens.vue';
 import ManagedAssetActions from '@/components/asset-manager/managed/ManagedAssetActions.vue';
-import ManagedAssetIgnoringMore from '@/components/asset-manager/managed/ManagedAssetIgnoringMore.vue';
+import ManagedAssetIgnoreSwitch from '@/components/asset-manager/managed/ManagedAssetIgnoreSwitch.vue';
 import DateDisplay from '@/components/display/DateDisplay.vue';
 import AssetDetailsBase from '@/components/helper/AssetDetailsBase.vue';
 import CopyButton from '@/components/helper/CopyButton.vue';
@@ -57,11 +57,13 @@ const deleteAsset = (asset: SupportedAsset) => emit('delete-asset', asset);
 const {
   isAssetWhitelisted,
   isSpamAsset,
+  loadingIgnore,
+  loadingSpam,
+  loadingWhitelist,
   massIgnore,
   toggleIgnoreAsset,
   toggleSpam,
   toggleWhitelistAsset,
-  useIsAssetIgnored,
 } = useManagedAssetOperations(() => emit('refresh'), ignoredFilter, selected);
 
 const { cols, data, expand, isExpanded } = useManagedAssetTable(
@@ -143,36 +145,18 @@ function getAssetLocation(row: SupportedAsset): string | undefined {
         {{ formatType(row.assetType) }}
       </template>
       <template #item.ignored="{ row }">
-        <div
+        <ManagedAssetIgnoreSwitch
           v-if="canBeIgnored(row)"
-          class="flex justify-start items-center gap-2"
-        >
-          <RuiTooltip
-            :popper="{ placement: 'top' }"
-            :open-delay="400"
-            tooltip-class="max-w-[10rem]"
-            :disabled="!isSpamAsset(row)"
-          >
-            <template #activator>
-              <RuiSwitch
-                color="primary"
-                hide-details
-                :disabled="isSpamAsset(row)"
-                :model-value="useIsAssetIgnored(row.identifier).value"
-                @update:model-value="toggleIgnoreAsset(row)"
-              />
-            </template>
-            {{ t('ignore.spam.hint') }}
-          </RuiTooltip>
-
-          <ManagedAssetIgnoringMore
-            v-if="showMoreOptions(row)"
-            :identifier="row.identifier"
-            :is-spam="isSpamAsset(row)"
-            @toggle-whitelist="toggleWhitelistAsset(row.identifier)"
-            @toggle-spam="toggleSpam(row)"
-          />
-        </div>
+          :identifier="row.identifier"
+          :is-spam="isSpamAsset(row)"
+          :show-more-options="showMoreOptions(row)"
+          :loading-ignore="loadingIgnore === row.identifier"
+          :loading-whitelist="loadingWhitelist === row.identifier"
+          :loading-spam="loadingSpam === row.identifier"
+          @toggle-ignore="toggleIgnoreAsset(row)"
+          @toggle-whitelist="toggleWhitelistAsset(row.identifier)"
+          @toggle-spam="toggleSpam(row)"
+        />
       </template>
       <template #item.actions="{ row }">
         <RowActions